requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.1
aioice==0.10.1
aiortc==1.14.0
aiosignal==1.4.0
alembic==1.17.0
annotated-doc==0.0.3
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
asteroid-filterbanks==0.4.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
audioread==3.0.1
av==16.0.1
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ci-sdr==0.0.2
click==8.3.0
coloredlogs==15.0.1
colorlog==6.10.1
configargparse==1.7.1
contourpy==1.3.2
cryptography==46.0.3
ctc-segmentation==1.7.4
ctranslate2==4.6.0
cycler==0.12.1
cython==3.1.6
decorator==5.2.1
deepface==0.0.95
distance==0.1.3
dnspython==2.8.0
docopt==0.6.2
einops==0.8.1
espnet==202205
espnet-model-zoo==0.1.7
espnet-tts-frontend==0.0.3
exceptiongroup==1.3.0
fast-bss-eval==0.1.3
fastapi==0.120.4
faster-whisper==1.2.0
filelock==3.20.0
fire==0.7.1
flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
g2p-en==2.1.0
gast==0.6.0
gdown==5.2.0
git-filter-repo==2.47.0
google-crc32c==1.7.1
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.76.0
gunicorn==23.0.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.1.10
huggingface-hub==0.36.0
humanfriendly==10.0
hyperpyyaml==1.2.2
idna==3.11
ifaddr==0.2.0
inflect==7.5.0
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jax==0.5.3
jaxlib==0.5.3
jinja2==3.1.6
joblib==1.5.2
julius==0.2.7
kaldiio==2.18.1
keras==3.11.3
kiwisolver==1.4.9
lazy-loader==0.4
libclang==18.1.1
librosa==0.11.0
lightning==2.5.5
lightning-utilities==0.15.2
llvmlite==0.45.1
lz4==4.4.4
mako==1.3.10
markdown==3.9
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.21
ml-dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
mtcnn==1.0.0
multidict==6.7.0
namex==0.1.0
networkx==3.4.2
nltk==3.9.2
numba==0.62.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnxruntime==1.23.2
openai-whisper==20250625
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt-einsum==3.4.0
optree==0.17.0
optuna==4.5.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pip==25.2
platformdirs==4.5.0
polars==1.35.1
polars-runtime-32==1.35.1
pooch==1.8.2
primepy==1.3
propcache==0.4.1
protobuf==4.25.8
psutil==7.1.3
pyannote-audio==3.4.0
pyannote-core==5.0.0
pyannote-database==5.1.3
pyannote-metrics==3.2.1
pyannote-pipeline==3.0.1
pycparser==2.23
pydantic==2.12.3
pydantic-core==2.41.4
pyee==13.0.0
pygments==2.19.2
pylibsrtp==1.0.0
pyopenssl==25.3.0
pyparsing==3.2.5
pypinyin==0.44.0
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==2.5.5
pytorch-metric-learning==2.9.0
pytorch-wpe==0.0.1
pytz==2025.2
pyworld==0.3.5
pyyaml==6.0.3
regex==2025.10.23
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
ruamel-yaml==0.18.16
ruamel-yaml-clib==0.2.14
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.15.3
semver==3.0.4
sentencepiece==0.2.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
speechbrain==1.0.3
sqlalchemy==2.0.44
starlette==0.49.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardx==2.6.4
tensorflow==2.19.1
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
tf-keras==2.19.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.2.0
torch-audiomentations==0.12.0
torch-complex==0.4.4
torch-pitch-shift==1.2.5
torchaudio==2.2.0
torchcodec==0.8.0
torchmetrics==1.8.2
torchvision==0.17.0
tqdm==4.67.1
transformers==4.57.1
triton==2.2.0
typeguard==4.4.4
typer==0.20.0
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
ultralytics==8.3.223
ultralytics-thop==2.0.18
unidecode==1.4.0
urllib3==2.5.0
uvicorn==0.38.0
werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.0
yarl==1.22.0
